Refactor / tidy up
